+++ backup/mumpsDocument.ts
@@ -5,18 +5,18 @@ export default class MumpsDocumentSymbolProvider implements vscode.DocumentSymbo
 
 	public provideDocumentSymbols(document: vscode.TextDocument): Promise<vscode.SymbolInformation[]> {
 		return new Promise(resolve => {
-			let parsedDoc = parser.parseText(document.getText());
-			let symbols: vscode.SymbolInformation[] = [];
+			const parsedDoc = parser.parseText(document.getText());
+			const symbols: vscode.SymbolInformation[] = [];
 
 			parsedDoc.methods.forEach(method => {
-				let kind = vscode.SymbolKind.Function;
+				const kind = vscode.SymbolKind.Function;
 
-				let startPosition = new vscode.Position(method.id.position.line, 0);
+				const startPosition = new vscode.Position(method.id.position.line, 0);
 
-				let endPostionLine = (method.endLine === -1) ? document.lineCount - 1 : method.endLine;
-				let endPosition = new vscode.Position(endPostionLine, 0);
+				const endPostionLine = (method.endLine === -1) ? document.lineCount - 1 : method.endLine;
+				const endPosition = new vscode.Position(endPostionLine, 0);
 
-				let methodRange = new vscode.Location(document.uri, new vscode.Range(startPosition, endPosition));
+				const methodRange = new vscode.Location(document.uri, new vscode.Range(startPosition, endPosition));
 
 				symbols.push(new vscode.SymbolInformation(method.id.value, kind, '', methodRange));
 			});